- PlotOptions now loads the PlotOptions section of config.yml.   Config.py now handles Schemes only (& will soon be renamed) - SuperGUI.py no longer depends on Config.py.
The idea here is to separate configuration & defaulting of plot
parameters (which is a concern of SuperGUI) from specifying the
appearance of plot elements (which is a concern of PlotLib).

For now both types of information live in config.yml, with the
former in the PlotOptions section and the latter in the Schemes
section.